// src/types/project.d.ts
export interface Project {
  title: string;
  description: string;
  fullDescription: string;
  technologies: string[];
  image: string;
  category: string;
  duration: string;
  team: string;
  challenges: string[];
  solutions: string[];
  results: string[];
  demoUrl: string;
  videoUrl: string;
  githubUrl: string;
  frontendGithubUrl?: string;
}